認知再構成法の基本的な機能を実装

+++ src/main/java/com/mentalapp/cbt_cr/entity/CbtCrPositiveFeel.java
@@ -6,9 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-/**
- * 認知再構成法とポジティブ感情の中間テーブルのエンティティクラス
- */
+/** 認知再構成法とポジティブ感情の中間テーブルのエンティティクラス */
 @Entity
 @Table(name = "cbt_cr_positive_feels")
 @Getter
@@ -16,22 +14,22 @@ import lombok.Setter;
 @NoArgsConstructor
 public class CbtCrPositiveFeel {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "id")
+  private Long id;
 
-    @Column(name = "cbt_cr_id", insertable = false, updatable = false)
-    private Long cbtCrId;
+  @Column(name = "cbt_cr_id", insertable = false, updatable = false)
+  private Long cbtCrId;
 
-    @Column(name = "positive_feel_id", insertable = false, updatable = false)
-    private Long positiveFeelId;
+  @Column(name = "positive_feel_id", insertable = false, updatable = false)
+  private Integer positiveFeelId;
 
-    @ManyToOne
-    @JoinColumn(name = "cbt_cr_id")
-    private CbtCr cbtCr;
+  @ManyToOne
+  @JoinColumn(name = "cbt_cr_id")
+  private CbtCr cbtCr;
 
-    @ManyToOne
-    @JoinColumn(name = "positive_feel_id")
-    private PositiveFeel positiveFeel;
+  @ManyToOne
+  @JoinColumn(name = "positive_feel_id")
+  private PositiveFeel positiveFeel;
 }
